--- ZoFo/GameCore/GameObjects/AttackEntities/Bullet.cs
@@ -1,7 +1,0 @@
-﻿using System;
-
-namespace ZoFo.GameCore.GameObjects;
-public class Bullet : Projectile
-{
-
-}
