--- ZoFo/GameCore/GameObjects/AttackEntities/Rock.cs
@@ -1,7 +1,0 @@
-﻿using System;
-
-namespace ZoFo.GameCore.GameObjects;
-public class Rock : Projectile
-{
-
-}
