--- ZoFo/GameCore/GameObjects/BaseClasses/Collectable.cs
@@ -1,7 +1,0 @@
-﻿using System;
-
-namespace ZoFo.GameCore.GameObjects;
-public class Collectable : Entity
-{
-
-}
