--- ZoFo/GameCore/GameObjects/BaseClasses/Entity.cs
@@ -1,33 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Graphics;
-
-namespace ZoFo.GameCore.GameObjects
-{
-    public abstract class Entity : GameObject
-    {
-        public int Id { get; set; }
-      public void CollisionComponent()
-        {
-
-        }
-
-        public void AnimationComponent()
-         {
-
-         }
-
-        public void UpdateLogic()
-        {
-
-        }
-
-    }
-}
-
-//вектор
-//вилосити
-//поситион
-//текстура
