--- ZoFo/GameCore/GameObjects/BaseClasses/LivingEntity.cs
@@ -1,20 +1,0 @@
-﻿using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Graphics;
-using System;
-
-namespace ZoFo.GameCore.GameObjects;
-public class LivingEntity : Entity
-{
-    public Vector2 velocity;
-    public Vector2 position;
-
-    public void TextureLoad(SpriteBatch spriteBatch)
-    {
-
-    }
-
-}
-
-
-
-
