--- ZoFo/GameCore/GameObjects/Enemy.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Net.Mime;
-using System.Reflection;
-using Microsoft.Xna.Framework.Content;
-
-namespace ZoFo.GameCore.GameObjects;
-public class Enemy : LivingEntity
-{
-
-}
