+++ ZoFo/GameCore/GameObjects/Entities/Entity.cs
@@ -1,0 +1,33 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Graphics;
+
+namespace ZoFo.GameCore.GameObjects.Entities
+{
+    public abstract class Entity : GameObject
+    {
+        public int Id { get; set; }
+        public void CollisionComponent()
+        {
+
+        }
+
+        public void AnimationComponent()
+        {
+
+        }
+
+        public void UpdateLogic()
+        {
+
+        }
+
+    }
+}
+
+//вектор
+//вилосити
+//поситион
+//текстура
