+++ ZoFo/GameCore/GameObjects/Entities/Interactables/Collectables/Collectable.cs
@@ -1,0 +1,7 @@
+﻿using System;
+
+namespace ZoFo.GameCore.GameObjects.Entities.Interactables.Collectables;
+public class Collectable : Entity
+{
+
+}
