+++ ZoFo/GameCore/GameObjects/Entities/LivingEntities/Enemies/Enemy.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Net.Mime;
+using System.Reflection;
+using Microsoft.Xna.Framework.Content;
+
+namespace ZoFo.GameCore.GameObjects.Entities.LivingEntities.Enemies;
+public class Enemy : LivingEntity
+{
+
+}
