+++ ZoFo/GameCore/GameObjects/Entities/LivingEntities/LivingEntity.cs
@@ -1,0 +1,21 @@
+﻿using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Graphics;
+using System;
+using ZoFo.GameCore.GameObjects.Entities;
+
+namespace ZoFo.GameCore.GameObjects.Entities.LivingEntities;
+public class LivingEntity : Entity
+{
+    public Vector2 velocity;
+    public Vector2 position;
+
+    public void TextureLoad(SpriteBatch spriteBatch)
+    {
+
+    }
+
+}
+
+
+
+
