+++ ZoFo/GameCore/GameObjects/Entities/LivingEntities/Player/Player.cs
@@ -1,0 +1,7 @@
+﻿using System;
+
+namespace ZoFo.GameCore.GameObjects.Entities.LivingEntities.Player;
+public class Player : LivingEntity
+{
+
+}
