+++ ZoFo/GameCore/GameObjects/Entities/LivingEntities/Player/PlayerAttacks/GunAttack.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ZoFo.GameCore.GameObjects
+namespace ZoFo.GameCore.GameObjects.Entities.LivingEntities.Player.PlayerAttacks
 {
-    internal interface IPlayerWeaponAttack
+    internal class GunAttack : IPlayerWeaponAttack
     {
 
     }
