+++ ZoFo/GameCore/GameObjects/Entities/LivingEntities/Player/PlayerAttacks/SwordAttack.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ZoFo.GameCore.GameObjects
+namespace ZoFo.GameCore.GameObjects.Entities.LivingEntities.Player.PlayerAttacks
 {
-    internal class SwordAttack:IPlayerWeaponAttack
+    internal class SwordAttack : IPlayerWeaponAttack
     {
 
     }
