+++ ZoFo/GameCore/GameObjects/Entities/LivingEntities/Projectiles/Rock.cs
@@ -1,0 +1,7 @@
+﻿using System;
+
+namespace ZoFo.GameCore.GameObjects.Entities.LivingEntities.Projectiles;
+public class Rock : Projectile
+{
+
+}
