--- ZoFo/GameCore/GameObjects/Player/Player.cs
@@ -1,7 +1,0 @@
-﻿using System;
-
-namespace ZoFo.GameCore.GameObjects;
-public class Player : LivingEntity
-{
-
-}
