+++ ZoFo/GameCore/GameObjects/Projectile.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using ZoFo.GameCore.GameObjects.Entities.LivingEntities;
 
 namespace ZoFo.GameCore.GameObjects;
 public class Projectile : LivingEntity
