+++ ZoFo/GameCore/GameObjects/StopObjects/StopObject.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace ZoFo.GameCore.GameObjects.StopObjects;
+
+public class StopObject
+{
+
+}
