--- ZoFo/GameCore/GameObjects/Tiles/StopObject.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace ZoFo.GameCore.GameObjects;
-
-public class StopObject
-{
-
-}
