+++ ZoFo/ZoFo.csproj
@@ -11,6 +11,11 @@
         <ApplicationIcon>Icon.ico</ApplicationIcon>
     </PropertyGroup>
     <ItemGroup>
+      <Compile Remove="GameCore\GameObjects\BaseClasses\**" />
+      <EmbeddedResource Remove="GameCore\GameObjects\BaseClasses\**" />
+      <None Remove="GameCore\GameObjects\BaseClasses\**" />
+    </ItemGroup>
+    <ItemGroup>
         <None Remove="Icon.ico" />
         <None Remove="Icon.bmp" />
     </ItemGroup>
